modified:   Dijkstra.cpp 	modified:   Dijkstra.h 	modified:   device.h 	modified:   dijkstra.cl 	modified:   main.cpp

+++ OpenCL/Dijkstra.h
@@ -3,7 +3,7 @@
 struct GraphData;
 
 void runDijkstraMultiGPUandCPU( cl_context gpuContext, cl_context cpuContext, GraphData* graph, 
-                                int *sourceVertices, long long *outResultCosts, int numResults);
+                                int *sourceVertices, int *outResultCosts, int numResults);
 
 void runDijkstraMultiGPU( cl_context gpuContext, GraphData* graph, int *sourceVertices,
-                          long long *outResultCosts, int numResults );
+                          int *outResultCosts, int numResults );
